Make inccounter a core contract
