+++ components/processors/component.go
@@ -1,13 +1,11 @@
 package processors
 
 import (
-	"go.uber.org/dig"
-
 	"github.com/iotaledger/hive.go/app"
-	"github.com/iotaledger/wasp/contracts/native/inccounter"
 	"github.com/iotaledger/wasp/packages/isc/coreutil"
 	"github.com/iotaledger/wasp/packages/vm/core/coreprocessors"
 	"github.com/iotaledger/wasp/packages/vm/processors"
+	"go.uber.org/dig"
 )
 
 func init() {
@@ -30,7 +28,7 @@ func provide(c *dig.Container) error {
 		Component.LogInfo("Registering native contracts...")
 
 		nativeContracts := []*coreutil.ContractProcessor{
-			inccounter.Processor,
+			// inccounter.Processor,
 		}
 
 		for _, c := range nativeContracts {
