+++ packages/vm/core/inccounter/state.go
